Put dashboard views into separate files
