+++ dashboard/urls.py
@@ -1,22 +1,25 @@
 from django.urls import path
 
-from . import views
+import dashboard.views.api_keys
+import dashboard.views.integrations
+import dashboard.views.stripe
+import dashboard.views.tasks
 
 urlpatterns = [
-    path("refill/", views.create_stripe_payment_link, name="refill_budget"),
+    path("refill/", dashboard.views.stripe.create_stripe_payment_link, name="refill_budget"),
     path(
         "add-slack-integration/",
-        views.add_slack_integration,
+        dashboard.views.integrations.add_slack_integration,
         name="add_slack_integration",
     ),
     path(
         "add-linear-integration/",
-        views.add_linear_integration,
+        dashboard.views.integrations.add_linear_integration,
         name="add_linear_integration",
     ),
-    path("tasks/", views.TaskListView.as_view(), name="task_list"),
-    path("integrations/", views.IntegrationView.as_view(), name="integrations"),
-    path("api-keys/", views.APIKeyListView.as_view(), name="api_key_list"),
-    path("tasks/<uuid:pk>/", views.TaskDetailView.as_view(), name="task_detail"),
-    path("tasks/<uuid:pk>/undo/", views.TaskUndoView.as_view(), name="task_undo"),
+    path("tasks/", dashboard.views.tasks.TaskListView.as_view(), name="task_list"),
+    path("integrations/", dashboard.views.integrations.IntegrationView.as_view(), name="integrations"),
+    path("api-keys/", dashboard.views.api_keys.APIKeyListView.as_view(), name="api_key_list"),
+    path("tasks/<uuid:pk>/", dashboard.views.tasks.TaskDetailView.as_view(), name="task_detail"),
+    path("tasks/<uuid:pk>/undo/", dashboard.views.tasks.TaskUndoView.as_view(), name="task_undo"),
 ]
